change product.context.jsx to categories.context.jsx and change code for loop categoriesMap as an Object Data Structure
